Update README.md
Combining feet/inches and meters conversions

Write a program that combines the functions from your previous solutions for Lab 19.3 and 19.24

The program should ask the user if they want to convert from feet and inches to meters or from meters to feet and inches. The program then performs the desired conversion. Have the user respond by typing the integer 1 for one type of conversion and 2 for the other conversion. The program reads the user’s answer and then executes an if-else statement.

Each branch of the if-else statement should be a function call. The two functions called in the if-else statement will have function definitions that are very similar to the programs for Lab 19.3 and 19.4. Thus, they will be function definitions that call other functions in their function bodies.

Important: Use function declaration/prototypes towards the top of the program for all of your programmer defined functions.

Optional:
Include a loop that lets the user repeat this computation for new input values until the user says he or she wants to end the program.